+++ ios-widget/TaleaWidget/TaleaWidgetBundle.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+import WidgetKit
+
+@main
+struct TaleaWidgetBundle: WidgetBundle {
+  var body: some Widget {
+    TaleaWidget()
+  }
+}